Add navigation to MainScreen
