+++ presentation/build.gradle.kts
@@ -58,6 +58,7 @@ dependencies {
     implementation("com.google.android.material:material:1.11.0")
     implementation(project(mapOf("path" to ":domain")))
     implementation(project(mapOf("path" to ":common")))
+    implementation("androidx.navigation:navigation-compose:2.7.6")
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.5")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
