+++ presentation/src/main/java/com/example/presentation/navigation/Screen.kt
@@ -1,0 +1,7 @@
+package com.example.presentation.navigation
+
+sealed class Screen(val route: String) {
+    object MainScreen : Screen("main_screen")
+
+    object DetailScreen : Screen("detail_screen")
+}
